chore: improve indexes
