+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export * from './GooglePayButton';
+export * from './ApplePayButton';
+export { AdyenCheckout } from './AdyenCheckout';
+export type { AdyenCheckoutProps } from './AdyenCheckout';
