--- src/core/AdyenCheckoutError.ts
@@ -1,35 +1,0 @@
-class AdyenCheckoutError extends Error {
-  protected static errorTypes = {
-    /** Network error. */
-    NETWORK_ERROR: 'NETWORK_ERROR',
-
-    /** Shopper canceled the current transaction. */
-    CANCEL: 'CANCEL',
-
-    /** Implementation error. The method or parameter are incorrect or are not supported. */
-    IMPLEMENTATION_ERROR: 'IMPLEMENTATION_ERROR',
-
-    /** Generic error. */
-    ERROR: 'ERROR',
-
-    /** Error while requesting new session from sessionData */
-    SESSION_ERROR: 'SESSION_ERROR',
-  };
-
-  public cause: unknown;
-
-  constructor(
-    type: keyof typeof AdyenCheckoutError.errorTypes,
-    message?: string,
-    options?: Error
-  ) {
-    super(message);
-
-    this.name = AdyenCheckoutError.errorTypes[type];
-
-    this.cause =
-      options instanceof Error ? options : new Error(String(options));
-  }
-}
-
-export default AdyenCheckoutError;
