+++ src/core/configurations/index.ts
