+++ src/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useAdyenCheckout } from './useAdyenCheckout';
+export type { AdyenCheckoutContextValue } from './useAdyenCheckout';
