+++ src/index.ts
@@ -1,45 +1,4 @@
-export * from './components/GooglePayButton';
-export * from './components/ApplePayButton';
-
-export { Card } from './core/types';
-export type * from './core/types';
-
-export { AdyenApplePay } from './modules/AdyenApplePay';
-export { AdyenGooglePay } from './modules/AdyenGooglePay';
-export { AdyenInstant } from './modules/AdyenInstant';
-
-export { AdyenDropIn } from './modules/DropInModule';
-export type { DropInModule } from './modules/DropInModule';
-
-export { AdyenCSE } from './modules/AdyenCSEModule';
-export type { AdyenCSEModule } from './modules/AdyenCSEModule';
-
-export { AdyenAction } from './modules/ActionModule';
-export type { ActionModule } from './modules/ActionModule';
-
-export type * from './core/AdyenNativeModules';
-
-export { useAdyenCheckout } from './hooks/useAdyenCheckout';
-export type { AdyenCheckoutContextType } from './hooks/useAdyenCheckout';
-
-export { AdyenCheckout } from './components/AdyenCheckout';
-export type { AdyenCheckoutProps } from './components/AdyenCheckout';
-
-export * from './core/constants';
-
-export type * from './core/configurations/Configuration';
-
-export type { PartialPaymentConfiguration } from './core/configurations/PartialPaymentConfiguration';
-export type { PartialPaymentComponent } from './wrappers/PartialPaymentsComponentWrapper';
-
-export type * from './core/configurations/GooglePayConfiguration';
-
-export type { DropInConfiguration } from './core/configurations/DropInConfiguration';
-
-export type { ThreeDSConfiguration } from './core/configurations/ThreeDSConfiguration';
-
-export type * from './core/configurations/AddressLookup';
-
-export type * from './core/configurations/ApplePayConfiguration';
-
-export type * from './core/configurations/CardsConfiguration';
+export * from './components';
+export * from './core';
+export * from './hooks';
+export * from './modules';
